fix: Improve IBAN guessing using BIC
French bank listing is not easy to obtain and schwifty does not have
guessing mechanisms to rebuild an IBAN from partial information. Use
existing BIC data and try to find a suitable entry in schwifty database
to build the IBAN.

This may result in incorrect IBAN and BIC being returned but it should be
sufficient to match transactions with beneficiary list and
transfer history.

Refs: https://github.com/mdomke/schwifty/issues/173
Refs: https://github.com/mdomke/schwifty/issues/242