fix: Export card number to memo
This information should already by available in card field, however this
does not appear to work consistently. Also, category does not really
related to memo but this is how backward compatibility is currently
implemented. This will make it possible to see which card was used for
the transaction in OFX export because not all cards appear to map well
to CCACCTINFO aggregate.